refactor(datagrid): split the grid coordinator's notification wiring into its own extension
